fix: always use cached connection, add test
